Working for loss of L, but just found bug on replace_subtree for C-loss
